Delete the none values to enable simplejson serialisation of Floats
